Move Bytes Command to Util Auth Module
