+++ petal/commands/util.py
@@ -20,6 +20,7 @@ from petal.exceptions import (
     CommandOperationError,
 )
 from petal.util import cdn, format
+from petal.util.bits import bytes_to_braille, chunk
 from petal.util.messages import member_message_history
 
 
@@ -680,6 +681,45 @@ class CommandsUtil(core.Commands):
                 )
 
         yield f"Showing last __{s}__ Messages."
+
+    async def cmd_bytes(self, src, **_):
+        """Encode the message provided into a Bytes object. Then, print it.
+
+        Debug utility to sanity check **__exactly__** what is received over Discord.
+
+        Syntax: `{p}bytes <literally anything>...`
+        """
+        raw: bytes = src.content[7:].encode("utf-8")
+
+        _bin: List[str] = [f"{b:0>8b}" for b in raw]
+        _hex: List[str] = [f"{b:0>2X}" for b in raw]
+
+        return (
+            discord.Embed(
+                title="Detailed String Analysis",
+                description=format.bold(format.escape(repr(raw)[2:-1])),
+                color=0xFFCD0A,
+            )
+            .add_field(
+                name="Binary",
+                value="\n".join(
+                    f"`{i * 4:0>2}`-`{min((i * 4 + 3, len(_bin) - 1)):0>2}` :: "
+                    + " ".join(f"__`{c}`__" for c in ch if c is not None)
+                    for i, ch in enumerate(chunk(_bin, 4))
+                ),
+            )
+            .add_field(
+                name="Hexadecimal",
+                value="\n".join(
+                    f"`{i * 16:0>2}`-`{min((i * 16 + 15, len(_hex) - 1)):0>2}` :: "
+                    + " ".join(f"__`{c}`__" for c in ch if c is not None)
+                    for i, ch in enumerate(chunk(_hex, 16))
+                ),
+            )
+            .add_field(
+                name="Raw Bits", value=format.bold(format.mono(bytes_to_braille(raw)))
+            )
+        )
 
 
 # Keep the actual classname unique from this common identifier
